[PULSEBS-118] renamed API.js to index.js
