+++ client/src/api/index.js
@@ -1,0 +1,7 @@
+import {isAuthenticated, userLogout, userLogin} from './authAPI' ;
+import {getBookings, bookLecture} from './bookingAPI' ;
+import {getLecture, getLectures} from './lectureAPI';
+
+const API = { userLogin, userLogout, isAuthenticated, getLectures, getLecture, getBookings, bookLecture }
+
+export default API
