+++ client/src/components/CalendarPage.js
@@ -7,7 +7,7 @@ import listPlugin from '@fullcalendar/list';
 import moment from 'moment';
 import { AuthContext } from '../auth/AuthContext';
 import CalendarModal from './CalendarModal';
-import API from '../api/API';
+import API from '../api';
 
 class CalendarPage extends React.Component {
   constructor(props) {
